requirements.txt
django-ip-geolocation
django-ratelimit